requirements.txt
google-genai>=1.0.0
markitdown[all]>=0.1.0
youtube-transcript-api>=0.6.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-multipart>=0.0.9
neo4j>=5.0.0
